feat: add JWT authentication with bcrypt and protected routes

+++ server/routes/foodRoutes.js
@@ -1,12 +1,13 @@
 import express from 'express'
 import pool from '../db/index.js'
+import authMiddleware from '../middleware/auth.js'
 
 const router = express.Router()
 
 // GET all foods → /api/foods
-router.get('/foods', async(req, res) => {
+router.get('/foods', authMiddleware, async(req, res) => {
     try{
-        const result = await pool.query('SELECT * FROM foods ORDER BY logged_at DESC')
+        const result = await pool.query('SELECT * FROM foods WHERE user_id = $1', [req.user.userId])
         res.json(result.rows)
     }
     catch(err){
@@ -16,10 +17,10 @@ router.get('/foods', async(req, res) => {
 })
 
 // GET single food by id → /api/foods/:id
-router.get('/foods/:id', async (req, res) => {
+router.get('/foods/:id', authMiddleware, async (req, res) => {
     try{
-        const id = req.params.id
-        const result = await pool.query('SELECT * FROM foods WHERE id = $1', [id])
+        const userId = req.user.userId 
+        const result = await pool.query('SELECT * FROM foods WHERE user_id = $1', [userId])
         res.json(result.rows[0]);
     }
     catch(err){
@@ -28,14 +29,14 @@ router.get('/foods/:id', async (req, res) => {
 })
 
 // POST - add new food → /api/foods
-router.post('/foods', async (req, res) =>{
+router.post('/foods',authMiddleware, async (req, res) =>{
     const {name, calories, protein} = req.body
     if (!name || !calories || !protein) {
         return res.status(400).json({ message: 'name, calories and protein are required' })
     }
 
     try{
-        const result = await pool.query('INSERT INTO foods (name, calories, protein) VALUES ($1, $2, $3) RETURNING *', [name, calories, protein])
+        const result = await pool.query('INSERT INTO foods (name, calories, protein, user_id) VALUES ($1, $2, $3, $4) RETURNING *', [name, calories, protein, req.user.userId])
         res.json(result.rows[0]);
     }
     catch(err){
@@ -44,10 +45,10 @@ router.post('/foods', async (req, res) =>{
 })
 
 // DELETE food by id → /api/foods/:id
-router.delete('/foods/:id', async (req, res) => {
+router.delete('/foods/:id',authMiddleware, async (req, res) => {
     try{
-        const id = req.params.id
-        const result = await pool.query('DELETE FROM foods WHERE id = $1', [id])
+        const userId = req.user.userId
+        const result = await pool.query('DELETE FROM foods WHERE id = $1 AND user_id = $2', [req.params.id, userId])
         res.status(200).json("Food is deleted")
     }
     catch(err){
